even more 64 bit updates

+++ UIKit/UIActionSheet+SA_Additions.m
@@ -15,7 +15,7 @@
 
 @implementation UIActionSheet (SA_AdditionsForButtons)
 
-- (void) addButtonWithTitle: (NSString *) title andSA_Tag: (int) tag {
+- (void) addButtonWithTitle: (NSString *) title andSA_Tag: (NSInteger) tag {
 	NSMutableDictionary				*dictionary = [self associatedValueForKey: kButtonTagsKey];
 	
 	if (dictionary == nil) {
@@ -37,7 +37,7 @@
 
 - (BOOL) shouldRunBlockAfterDismissal { return [[self associatedValueForKey: kRunBlockAfterDismissKey] boolValue]; }
 
-- (int) SA_TagForButtonAtIndex: (NSUInteger) index {
+- (NSInteger) SA_TagForButtonAtIndex: (NSUInteger) index {
 	if (index >= self.numberOfButtons) return 0;
 	NSMutableDictionary				*dictionary = [self associatedValueForKey: kButtonTagsKey];
